Maxflow

+++ src/City.cpp
@@ -1,8 +1,8 @@
-#include "City.h"
+#include "h/City.h"
 
 using namespace std;
 
-City::City(string name, int id, string code, int demand, int population) {
+City::City(string name, int id, string code, double demand, int population) {
     this->city_name = name;
     this->city_id = id;
     this->city_code = code;
@@ -19,8 +19,8 @@ void City::setCityId(int id) {this->city_id = id;}
 string City::getCityCode() const {return city_code;}
 void City::setCityCode(string code) {this->city_code = code;}
 
-int City::getCityDemand() const {return city_demand;}
-void City::setCityDemand(int demand) {this->city_demand = demand;}
+double City::getCityDemand() const {return city_demand;}
+void City::setCityDemand(double demand) {this->city_demand = demand;}
 
 int City::getCityPopulation() const {return city_population;}
 void City::setCityPopulation(int pop) {this->city_population = pop;}
